requirements.txt
asgiref==3.8.1
Django==4.0
gunicorn==23.0.0
packaging==24.2
python-dotenv==1.0.1
sqlparse==0.5.1
typing_extensions==4.12.2
whitenoise==6.8.2
